+++ sources/server/web/rest/service/AutomationRule.cpp
@@ -112,7 +112,7 @@ namespace web
             }
             catch (CRuleException& e)
             {
-               return CResult::GenerateError(std::string("Fail to retreive rule : ") + e.what());
+               return CResult::GenerateError(std::string("Fail to retrieve rule : ") + e.what());
             }
             catch (std::exception& ex)
             {
@@ -120,7 +120,7 @@ namespace web
             }
             catch (...)
             {
-               return CResult::GenerateError("unknown exception in retreiving the rule");
+               return CResult::GenerateError("unknown exception in retrieving the rule");
             }
          }
 
@@ -138,7 +138,7 @@ namespace web
             }
             catch (CRuleException& e)
             {
-               return CResult::GenerateError(std::string("Fail to retreive rule code : ") + e.what());
+               return CResult::GenerateError(std::string("Fail to retrieve rule code : ") + e.what());
             }
             catch (std::exception& ex)
             {
@@ -146,7 +146,7 @@ namespace web
             }
             catch (...)
             {
-               return CResult::GenerateError("unknown exception in retreiving the rule");
+               return CResult::GenerateError("unknown exception in retrieving the rule");
             }
          }
 
@@ -164,7 +164,7 @@ namespace web
             }
             catch (CRuleException& e)
             {
-               return CResult::GenerateError(std::string("Fail to retreive rule code : ") + e.what());
+               return CResult::GenerateError(std::string("Fail to retrieve rule code : ") + e.what());
             }
             catch (std::exception& ex)
             {
@@ -172,7 +172,7 @@ namespace web
             }
             catch (...)
             {
-               return CResult::GenerateError("unknown exception in retreiving the rule");
+               return CResult::GenerateError("unknown exception in retrieving the rule");
             }
          }
 
@@ -190,7 +190,7 @@ namespace web
             }
             catch (CRuleException& e)
             {
-               return CResult::GenerateError(std::string("Fail to retreive rule code : ") + e.what());
+               return CResult::GenerateError(std::string("Fail to retrieve rule code : ") + e.what());
             }
             catch (std::exception& ex)
             {
@@ -198,7 +198,7 @@ namespace web
             }
             catch (...)
             {
-               return CResult::GenerateError("unknown exception in retreiving the rule");
+               return CResult::GenerateError("unknown exception in retrieving the rule");
             }
          }
 
@@ -215,7 +215,7 @@ namespace web
             }
             catch (CRuleException& e)
             {
-               return CResult::GenerateError(std::string("Fail to retreive rule code : ") + e.what());
+               return CResult::GenerateError(std::string("Fail to retrieve rule code : ") + e.what());
             }
             catch (std::exception& ex)
             {
@@ -223,7 +223,7 @@ namespace web
             }
             catch (...)
             {
-               return CResult::GenerateError("unknown exception in retreiving the rule");
+               return CResult::GenerateError("unknown exception in retrieving the rule");
             }
          }