+++ sources/server/web/rest/service/Device.cpp
@@ -4,6 +4,7 @@
 #include "web/rest/Result.h"
 #include "web/rest/RestDispatcherHelpers.hpp"
 #include "communication/callback/SynchronousCallback.h"
+#include "database/common/requesters/Keyword.h"
 
 namespace web
 {
@@ -20,15 +21,13 @@ namespace web
                           boost::shared_ptr<dataAccessLayer::IKeywordManager> keywordManager,
                           communication::ISendMessageAsync& messageSender)
             : m_dataProvider(dataProvider),
+              m_deviceRequester(dataProvider->getDeviceRequester()),
               m_pluginManager(pluginManager),
               m_deviceManager(deviceManager),
               m_keywordManager(keywordManager),
               m_messageSender(messageSender)
          {
          }
-
-
-         CDevice::~CDevice() = default;
 
          const std::string& CDevice::getRestKeyword()
          {
@@ -39,6 +38,7 @@ namespace web
          {
             REGISTER_DISPATCHER_HANDLER(dispatcher, "GET", (m_restKeyword), CDevice::getAllDevices);
             REGISTER_DISPATCHER_HANDLER(dispatcher, "GET", (m_restKeyword)("*"), CDevice::getOneDevice);
+            REGISTER_DISPATCHER_HANDLER(dispatcher, "GET", (m_restKeyword)("*")("compatibleForMergeDevice"), CDevice::getCompatibleForMergeDevice);
             REGISTER_DISPATCHER_HANDLER(dispatcher, "GET", (m_restKeyword)("*")("configurationSchema"), CDevice::getDeviceConfigurationSchema);
             REGISTER_DISPATCHER_HANDLER(dispatcher, "GET", (m_restKeyword)("keyword"), CDevice::getAllKeywords);
             REGISTER_DISPATCHER_HANDLER(dispatcher, "GET", (m_restKeyword)("keyword")("*"), CDevice::getKeyword);
@@ -49,31 +49,121 @@ namespace web
             REGISTER_DISPATCHER_HANDLER(dispatcher, "GET", (m_restKeyword)("*")("keyword"), CDevice::getDeviceKeywords);
             REGISTER_DISPATCHER_HANDLER_WITH_INDIRECTOR(dispatcher, "PUT", (m_restKeyword)("*")("configuration"), CDevice::updateDeviceConfiguration,
                CDevice::transactionalMethod);
-            REGISTER_DISPATCHER_HANDLER_WITH_INDIRECTOR(dispatcher, "PUT", (m_restKeyword)("*")("restore"), CDevice::restoreDevice, CDevice::
-               transactionalMethod);
+            REGISTER_DISPATCHER_HANDLER_WITH_INDIRECTOR(dispatcher, "PUT", (m_restKeyword)("merge"), CDevice::mergeDevices,
+               CDevice:: transactionalMethod);
+            REGISTER_DISPATCHER_HANDLER_WITH_INDIRECTOR(dispatcher, "PUT", (m_restKeyword)("*")("restore"), CDevice::restoreDevice,
+               CDevice:: transactionalMethod);
             REGISTER_DISPATCHER_HANDLER_WITH_INDIRECTOR(dispatcher, "PUT", (m_restKeyword)("keyword")("*"), CDevice::updateKeywordFriendlyName,
                CDevice::transactionalMethod);
-            REGISTER_DISPATCHER_HANDLER_WITH_INDIRECTOR(dispatcher, "PUT", (m_restKeyword)("keyword")("*")("blacklist"), CDevice::
-               updateKeywordBlacklist, CDevice::transactionalMethod);
+            REGISTER_DISPATCHER_HANDLER_WITH_INDIRECTOR(dispatcher, "PUT", (m_restKeyword)("keyword")("*")("blacklist"),
+               CDevice:: updateKeywordBlacklist, CDevice::transactionalMethod);
             REGISTER_DISPATCHER_HANDLER_WITH_INDIRECTOR(dispatcher, "POST", (m_restKeyword)("keyword")("*")("command"), CDevice::sendKeywordCommand,
                CDevice::transactionalMethod);
-            REGISTER_DISPATCHER_HANDLER_WITH_INDIRECTOR(dispatcher, "DELETE", (m_restKeyword)("*")("*"), CDevice::deleteDevice, CDevice::
-               transactionalMethod);
+            REGISTER_DISPATCHER_HANDLER_WITH_INDIRECTOR(dispatcher, "DELETE", (m_restKeyword)("*")("*"), CDevice::deleteDevice,
+               CDevice:: transactionalMethod);
          }
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getOneDevice(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getOneDevice(const std::vector<std::string>& parameters,
+                                                                                           const std::string& requestContent) const
          {
             std::string objectId;
             if (parameters.size() > 1)
                objectId = parameters[1];
 
-            const auto deviceFound = m_dataProvider->getDeviceRequester()->getDevice(boost::lexical_cast<int>(objectId), true);
+            const auto deviceFound = m_deviceRequester->getDevice(boost::lexical_cast<int>(objectId), true);
             return CResult::GenerateSuccess(deviceFound);
          }
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getDeviceConfigurationSchema(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getCompatibleForMergeDevice(const std::vector<std::string>& parameters,
+                                                                                                          const std::string& requestContent) const
+         {
+            try
+            {
+               if (parameters.size() <= 1)
+                  return CResult::GenerateError("invalid parameter. Can not retrieve device id in url");
+
+               const auto refDeviceId = boost::lexical_cast<int>(parameters[1]);
+
+               // A device is compatible for merge with an other when :
+               // - plugins are compatible (see m_deviceRequester->getCompatibleForMergeDevice)
+               // - They contain at least one compatible keyword. Keywords are compatible when :
+               //   - capacityName, accessMode, name, type, units, typeInfo, measure and details are the same
+
+               // Get all devices with same characteristics of refDevice
+               auto compatibleDevices = m_deviceRequester->getCompatibleForMergeDevice(refDeviceId);
+               shared::CDataContainer commonKeywordsForCompatibleDevices;
+               if (!compatibleDevices.empty())
+               {
+                  // Second filter, check keywords from already filtered devices
+                  const auto refKeywords = m_keywordManager->getKeywords(refDeviceId);
+                  // Remove device from compatibles devices list if no common keyword exist between reference and first-filtered device
+                  compatibleDevices.erase(
+                     std::remove_if(compatibleDevices.begin(),
+                                    compatibleDevices.end(),
+                                    [this, &refKeywords, &commonKeywordsForCompatibleDevices](
+                                    const boost::shared_ptr<database::entities::CDevice>& candidateDevice)
+                                    {
+                                       auto candidateKeywords = m_keywordManager->getKeywords(candidateDevice->Id);
+                                       std::vector<shared::CDataContainer> commonKeywords;
+                                       // Iterate through reference keywords to find a common keyword in candidateDevice
+                                       for (const auto& refKeyword : refKeywords)
+                                       {
+                                          // Iterate through candidate keywords and remove consumed candidate to not used it more than one time
+                                          candidateKeywords.erase(
+                                             std::remove_if(candidateKeywords.begin(),
+                                                            candidateKeywords.end(),
+                                                            [&commonKeywords, &candidateDevice, &refKeyword](
+                                                            const boost::shared_ptr<database::entities::CKeyword>& candidateKeyword)
+                                                            {
+                                                               if (refKeyword->CapacityName == candidateKeyword->CapacityName &&
+                                                                  refKeyword->AccessMode == candidateKeyword->AccessMode &&
+                                                                  refKeyword->Name == candidateKeyword->Name &&
+                                                                  refKeyword->Type == candidateKeyword->Type &&
+                                                                  refKeyword->Units == candidateKeyword->Units &&
+                                                                  refKeyword->TypeInfo == candidateKeyword->TypeInfo &&
+                                                                  refKeyword->Measure == candidateKeyword->Measure &&
+                                                                  refKeyword->Details == candidateKeyword->Details)
+                                                               {
+                                                                  // A common device was found
+                                                                  shared::CDataContainer commonKeyword;
+                                                                  commonKeyword.set("from", refKeyword);
+                                                                  commonKeyword.set("to", candidateKeyword);
+                                                                  commonKeywords.push_back(commonKeyword);
+                                                                  return true;
+                                                               }
+                                                               return false;
+                                                            }),
+                                             candidateKeywords.end());
+                                       }
+
+                                       if (commonKeywords.empty())
+                                       {
+                                          // Remove device from compatible devices list if no common keyword
+                                          return true;
+                                       }
+
+                                       commonKeywordsForCompatibleDevices.set(std::to_string(candidateDevice->Id), commonKeywords);
+                                       return false;
+                                    }),
+                     compatibleDevices.end());
+               }
+               shared::CDataContainer collection;
+               collection.set("compatibleDevices", compatibleDevices);
+               collection.set("commonKeywordsByDevice", commonKeywordsForCompatibleDevices);
+               return CResult::GenerateSuccess(collection);
+            }
+            catch (std::exception& ex)
+            {
+               return CResult::GenerateError(ex);
+            }
+            catch (...)
+            {
+               return CResult::GenerateError("unknown exception in executing getDeviceConfigurationSchema query");
+            }
+         }
+
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getDeviceConfigurationSchema(const std::vector<std::string>& parameters,
+                                                                                                           const std::string& requestContent) const
          {
             try
             {
@@ -124,7 +214,7 @@ namespace web
          boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getAllDevices(
             const std::vector<std::string>& parameters, const std::string& requestContent) const
          {
-            const auto dvList = m_dataProvider->getDeviceRequester()->getDevices(true);
+            const auto dvList = m_deviceRequester->getDevices(true);
             shared::CDataContainer collection;
             collection.set(getRestKeyword(), dvList);
             return CResult::GenerateSuccess(collection);
@@ -154,8 +244,8 @@ namespace web
             }
          }
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getAllKeywords(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getAllKeywords(const std::vector<std::string>& parameters,
+                                                                                             const std::string& requestContent) const
          {
             try
             {
@@ -174,8 +264,8 @@ namespace web
             }
          }
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getDevicesWithCapacity(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getDevicesWithCapacity(const std::vector<std::string>& parameters,
+                                                                                                     const std::string& requestContent) const
          {
             try
             {
@@ -186,7 +276,7 @@ namespace web
                   const auto capacityName = parameters[3];
 
                   //run query
-                  const auto result = m_dataProvider->getDeviceRequester()->getDeviceWithCapacity(capacityName, cam);
+                  const auto result = m_deviceRequester->getDeviceWithCapacity(capacityName, cam);
                   shared::CDataContainer collection;
                   collection.set(getRestKeyword(), result);
                   return CResult::GenerateSuccess(collection);
@@ -203,8 +293,8 @@ namespace web
             }
          }
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getDeviceWithCapacityType(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getDeviceWithCapacityType(const std::vector<std::string>& parameters,
+                                                                                                        const std::string& requestContent) const
          {
             try
             {
@@ -214,7 +304,7 @@ namespace web
                   const shared::plugin::yPluginApi::EKeywordDataType typ(parameters[3]);
 
                   //run query
-                  const auto result = m_dataProvider->getDeviceRequester()->getDeviceWithCapacityType(cam, typ);
+                  const auto result = m_deviceRequester->getDeviceWithCapacityType(cam, typ);
                   shared::CDataContainer collection;
                   collection.set(getRestKeyword(), result);
                   return CResult::GenerateSuccess(collection);
@@ -232,7 +322,8 @@ namespace web
          }
 
          boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getDeviceWithKeywordAccessMode(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+            const std::vector<std::string>& parameters,
+            const std::string& requestContent) const
          {
             try
             {
@@ -241,12 +332,12 @@ namespace web
                   const shared::plugin::yPluginApi::EKeywordAccessMode cam(parameters[2]);
 
                   //run query
-                  const auto result = m_dataProvider->getDeviceRequester()->getDeviceWithKeywordAccessMode(cam);
+                  const auto result = m_deviceRequester->getDeviceWithKeywordAccessMode(cam);
                   shared::CDataContainer collection;
                   collection.set(getRestKeyword(), result);
                   return CResult::GenerateSuccess(collection);
                }
-               return CResult::GenerateError("invalid parameter. Can not retrieve access mode in url");
+               return CResult::GenerateError("invalid parameter. Can not retrieve access-mode in url");
             }
             catch (std::exception& ex)
             {
@@ -259,8 +350,8 @@ namespace web
          }
 
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getDeviceKeywordsForCapacity(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getDeviceKeywordsForCapacity(const std::vector<std::string>& parameters,
+                                                                                                           const std::string& requestContent) const
          {
             try
             {
@@ -288,15 +379,15 @@ namespace web
             }
          }
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getDeviceKeywords(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::getDeviceKeywords(const std::vector<std::string>& parameters,
+                                                                                                const std::string& requestContent) const
          {
             try
             {
                if (parameters.size() > 1)
                {
                   const auto deviceId = boost::lexical_cast<int>(parameters[1]);
-                  const auto deviceInDatabase = m_dataProvider->getDeviceRequester()->getDevice(deviceId, true);
+                  const auto deviceInDatabase = m_deviceRequester->getDevice(deviceId, true);
 
                   if (deviceInDatabase)
                   {
@@ -319,8 +410,8 @@ namespace web
             }
          }
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::sendKeywordCommand(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::sendKeywordCommand(const std::vector<std::string>& parameters,
+                                                                                                 const std::string& requestContent) const
          {
             try
             {
@@ -351,8 +442,8 @@ namespace web
             }
          }
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::deleteDevice(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::deleteDevice(const std::vector<std::string>& parameters,
+                                                                                           const std::string& requestContent) const
          {
             try
             {
@@ -389,8 +480,8 @@ namespace web
          }
 
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::updateDeviceFriendlyName(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::updateDeviceFriendlyName(const std::vector<std::string>& parameters,
+                                                                                                       const std::string& requestContent) const
          {
             try
             {
@@ -404,10 +495,10 @@ namespace web
                   if (deviceToUpdate.FriendlyName.isDefined())
                   {
                      //update data in base
-                     m_dataProvider->getDeviceRequester()->updateDeviceFriendlyName(deviceId, deviceToUpdate.FriendlyName());
+                     m_deviceRequester->updateDeviceFriendlyName(deviceId, deviceToUpdate.FriendlyName());
 
                      //return the device info
-                     const auto deviceFound = m_dataProvider->getDeviceRequester()->getDevice(deviceId, true);
+                     const auto deviceFound = m_deviceRequester->getDevice(deviceId, true);
                      return CResult::GenerateSuccess(deviceFound);
                   }
                   return CResult::GenerateError("invalid request content. could not retrieve device friendlyName");
@@ -425,8 +516,8 @@ namespace web
          }
 
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::updateDeviceConfiguration(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::updateDeviceConfiguration(const std::vector<std::string>& parameters,
+                                                                                                        const std::string& requestContent) const
          {
             try
             {
@@ -442,24 +533,24 @@ namespace web
                   //update friendly name
                   if (deviceToUpdate.FriendlyName.isDefined())
                   {
-                     m_dataProvider->getDeviceRequester()->updateDeviceFriendlyName(deviceId, deviceToUpdate.FriendlyName());
+                     m_deviceRequester->updateDeviceFriendlyName(deviceId, deviceToUpdate.FriendlyName());
                   }
 
                   if (deviceToUpdate.Model.isDefined())
                   {
-                     m_dataProvider->getDeviceRequester()->updateDeviceModel(deviceId, deviceToUpdate.Model());
+                     m_deviceRequester->updateDeviceModel(deviceId, deviceToUpdate.Model());
                   }
 
                   if (deviceToUpdate.Configuration.isDefined())
                   {
                      //update data in base
-                     m_dataProvider->getDeviceRequester()->updateDeviceConfiguration(deviceId, deviceToUpdate.Configuration());
+                     m_deviceRequester->updateDeviceConfiguration(deviceId, deviceToUpdate.Configuration());
                      m_messageSender.sendSetDeviceConfiguration(deviceId,
                                                                 deviceToUpdate.Configuration());
                   }
 
                   //return the device info
-                  const auto deviceFound = m_dataProvider->getDeviceRequester()->getDevice(deviceId, true);
+                  const auto deviceFound = m_deviceRequester->getDevice(deviceId, true);
                   return CResult::GenerateSuccess(deviceFound);
                }
                return CResult::GenerateError("invalid parameter. Can not retrieve device id in url");
@@ -475,8 +566,58 @@ namespace web
          }
 
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::restoreDevice(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::mergeDevices(const std::vector<std::string>& parameters,
+                                                                                           const std::string& requestContent) const
+         {
+            try
+            {
+               shared::CDataContainer content(requestContent);
+               const auto sourceDeviceId = content.get<int>("sourceDeviceId");
+               const auto targetDeviceId = content.get<int>("targetDeviceId");
+               const auto keywordCorrespondences = content.get<std::vector<shared::CDataContainer>>("keywordCorrespondences");
+
+               // Merge acquisitions
+               YADOMS_LOG(information) << "Starting merging 2 devices...";
+               for (const auto& keywordCorrespondence : keywordCorrespondences)
+               {
+                  // Move acquisitions
+                  const auto fromKw = keywordCorrespondence.get<int>("from");
+                  const auto toKw = keywordCorrespondence.get<int>("to");
+                  m_dataProvider->getAcquisitionRequester()->moveAllData(fromKw, toKw);
+                  // Update last acquisition in keyword table
+                  const auto lastData = m_dataProvider->getAcquisitionRequester()->getKeywordData(toKw,
+                                                                                                  boost::posix_time::not_a_date_time,
+                                                                                                  boost::posix_time::not_a_date_time,
+                                                                                                  1);
+                  m_dataProvider->getKeywordRequester()->updateLastValue(fromKw,
+                                                                         lastData.size() != 0
+                                                                            ? lastData[0].get<0>()
+                                                                            : boost::posix_time::not_a_date_time,
+                                                                         lastData.size() != 0 ? lastData[0].get<1>() : std::string());
+               }
+               // Change name of target device, to make plugin using this device from now
+               m_deviceRequester->rename(targetDeviceId, m_deviceRequester->getDevice(sourceDeviceId)->Name());
+
+               // Remove source device
+               m_deviceRequester->removeDevice(sourceDeviceId);
+
+               YADOMS_LOG(information) << "Device merge done...";
+
+               return CResult::GenerateSuccess();
+            }
+            catch (std::exception& ex)
+            {
+               return CResult::GenerateError(ex);
+            }
+            catch (...)
+            {
+               return CResult::GenerateError("unknown exception in merging devices");
+            }
+         }
+
+
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::restoreDevice(const std::vector<std::string>& parameters,
+                                                                                            const std::string& requestContent) const
          {
             try
             {
@@ -486,7 +627,7 @@ namespace web
                   m_deviceManager->updateDeviceBlacklistState(deviceId, false);
 
                   //return the device info
-                  const auto deviceFound = m_dataProvider->getDeviceRequester()->getDevice(deviceId, true);
+                  const auto deviceFound = m_deviceRequester->getDevice(deviceId, true);
                   return CResult::GenerateSuccess(deviceFound);
                }
                return CResult::GenerateError("invalid parameter. Can not retrieve device id in url");
@@ -502,8 +643,8 @@ namespace web
          }
 
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::updateKeywordFriendlyName(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::updateKeywordFriendlyName(const std::vector<std::string>& parameters,
+                                                                                                        const std::string& requestContent) const
          {
             try
             {
@@ -533,8 +674,8 @@ namespace web
             }
          }
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::updateKeywordBlacklist(
-            const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::updateKeywordBlacklist(const std::vector<std::string>& parameters,
+                                                                                                     const std::string& requestContent) const
          {
             try
             {
@@ -565,8 +706,9 @@ namespace web
          }
 
 
-         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::transactionalMethod(
-            CRestDispatcher::CRestMethodHandler realMethod, const std::vector<std::string>& parameters, const std::string& requestContent) const
+         boost::shared_ptr<shared::serialization::IDataSerializable> CDevice::transactionalMethod(CRestDispatcher::CRestMethodHandler realMethod,
+                                                                                                  const std::vector<std::string>& parameters,
+                                                                                                  const std::string& requestContent) const
          {
             auto pTransactionalEngine = m_dataProvider->getTransactionalEngine();
             boost::shared_ptr<shared::serialization::IDataSerializable> result;