+++ sources/server/changelog.md
@@ -1,4 +1,5 @@
 ### 2.2.0-beta.4
+* New feature #513 : Add device fusion 
 * Fix script interpreter package generation (post build copy)
 * Fix #529 : Fix saving system configuration
 * Fix #527 - Equipment renamed is not displayed immediately